math with floats

+++ src/vehicle.cpp
@@ -2774,8 +2774,8 @@ float vehicle::fuel_specific_energy( const itype_id &ftype ) const
     for( auto vehicle_part : parts ) {
         if( vehicle_part.is_tank() && vehicle_part.ammo_current() == ftype  &&
             vehicle_part.base.contents_made_of( LIQUID ) ) {
-            int energy = vehicle_part.base.contents.front().specific_energy;
-            int mass = to_gram( vehicle_part.base.contents.front().weight() );
+            float energy = vehicle_part.base.contents.front().specific_energy;
+            float mass = to_gram( vehicle_part.base.contents.front().weight() );
             total_energy += energy * mass;
             total_mass += mass;
         }
